+++ .prettierignore
@@ -1,5 +1,6 @@
 package.json
-package-lock.json
 .sanity
 .next
 dist
+
+node_modules
