+++ app/(frontend)/globals.css
