+++ app/(sanity)/admin/[[...index]]/page.tsx
@@ -1,0 +1,9 @@
+'use client';
+
+import { NextStudio } from 'next-sanity/studio';
+
+import config from '@/sanity.config';
+
+export default function AdminPage() {
+  return <NextStudio config={config} />;
+}
