--- sanity/.eslintrc
@@ -1,4 +1,0 @@
-{
-  "root": true,
-  "extends": "@sanity/eslint-config-studio"
-}
