+++ sanity/client.ts
@@ -1,0 +1,5 @@
+import { createClient } from 'next-sanity';
+
+import config from './config';
+
+export const client = createClient(config);
