+++ sanity/config.ts
@@ -1,8 +1,8 @@
-import { createClient } from 'next-sanity';
-
-export const client = createClient({
+const config = {
   projectId: 'q7v95rg2',
   dataset: 'production',
   useCdn: false,
   apiVersion: '2021-10-21',
-});
+};
+
+export default config;
