--- sanity/package.json
@@ -1,39 +1,0 @@
-{
-  "name": "mvp-nextjs-sanity",
-  "private": true,
-  "version": "1.0.0",
-  "main": "package.json",
-  "license": "UNLICENSED",
-  "scripts": {
-    "dev": "sanity dev",
-    "start": "sanity start",
-    "build": "sanity build",
-    "deploy": "sanity deploy",
-    "deploy-graphql": "sanity graphql deploy"
-  },
-  "keywords": [
-    "sanity"
-  ],
-  "dependencies": {
-    "@sanity/vision": "^3.16.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-icons": "^4.11.0",
-    "react-is": "^18.2.0",
-    "sanity": "^3.16.0",
-    "sanity-plugin-media": "^2.2.2",
-    "sanity-plugin-simpler-color-input": "^1.1.0",
-    "styled-components": "^5.3.9"
-  },
-  "devDependencies": {
-    "@sanity/eslint-config-studio": "^3.0.1",
-    "@types/react": "^18.0.25",
-    "@types/styled-components": "^5.1.26",
-    "eslint": "^8.6.0",
-    "prettier": "^3.0.2",
-    "typescript": "^5.1.6"
-  },
-  "engines": {
-    "pnpm": ">=8"
-  }
-}
