--- sanity/sanity.cli.ts
@@ -1,8 +1,0 @@
-import { defineCliConfig } from 'sanity/cli';
-
-export default defineCliConfig({
-  api: {
-    projectId: 'q7v95rg2',
-    dataset: 'production',
-  },
-});
